ElectricVehiclesR: Examining Electric Vehicle Data in Washington State, USA.

This package includes an RShiny app that analyzes electric vehicles in the State of Washington, USA. The app allows users to test combinations of variables and reach their own conclusions regarding electric vehicles.

Details

The package contains the following functions:

filter_vehicles: Filters the dataset based on electric range, make, and vehicle type.

summary_filtered_data: Generates summary statistics for the filtered data.

load_app: Launches the Shiny app to interactively explore electric vehicle options.

Data Source

The data was extracted from Kaggle and is titled "Electric Vehicle Population". You can access it here.

See also

Author

Maintainer: Rayyan Aamir raam0001@student.monash.edu