Electric Vehicle Data

Electric Vehicle Data

Usage

clean_vehicle

Format

'clean_vehicle'

A data frame with 181038 rows and 13 columns

county

Name of county from where the data is gathered

city

Name of city from where the data is gathered

state

Name of state from where the data is gathered

postal_code

The postal code from where the data is present

model_year

Manufacturing year of the model mentioned in the dataset

make

Manufacturer of the vehicle

model

Model Name of the mentioned vehicle

electric_vehicle_type

Type of the electric vehicle present in the dataset

clean_alternative_fuel_vehicle_cafv_eligibility

Clean Alternative Fuel Vehicle (CAFV) eligibility for the data present in this dataset

electric_range

The electric range the car provides

vehicle_location

Vehicle coordinates

lon

Longitude of the vehicle's location

lat

Latitude of the vehicle's location

Source

https://www.kaggle.com/datasets/jainaru/electric-vehicle-population/data